pyproject.toml
[project]
name = "hyperliquid-telegram-bot"
version = "0.1.0"
description = "-"
readme = "README.md"
requires-python = ">=3.10, < 3.11"
dependencies = [
    "hyperliquid-python-sdk>=0.7.1",
    "matplotlib>=3.9.2",
    "mplfinance>=0.12.10b0",
    "pandas-stubs>=2.2.3.241126",
    "pandas-ta",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.5",
    "pytest-mock>=3.12.0",
    "python-telegram-bot[job-queue]>=21.6",
    "setuptools>=75.5.0",
    "tabulate>=0.9.0",
    "types-requests>=2.32.0.20241016",
    "types-tabulate>=0.9.0.20240106",
    "tzlocal>=5.2",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
pythonpath = ["."]
filterwarnings = [
    "ignore:pkg_resources is deprecated as an API:DeprecationWarning"
]

[[tool.mypy.overrides]]
module = "hyperliquid.utils.types"
ignore_missing_imports = true
[[tool.mypy.overrides]]
module = "hyperliquid.exchange"
ignore_missing_imports = true
[[tool.mypy.overrides]]
module = "hyperliquid.info"
ignore_missing_imports = true
[[tool.mypy.overrides]]
module = "hyperliquid.utils"
ignore_missing_imports = true

[tool.uv.sources]
pandas-ta = { git = "https://github.com/twopirllc/pandas-ta.git", branch = "development" }
